Seperate unittests into logical units
As well as making the code a little easier to navigate,
this allows for running on unit and only running the next
if the first passes. This is usefull in cases where some functionality
in the core library may be broken, making it easier to see the core error
and making the tests faster to run.